# openapi.json
{"openapi":"3.1.0","info":{"title":"Meeting Transcriber API","description":"\n    API pour la transcription de réunions audio en texte.\n    \n    Cette API permet :\n    - L'inscription et l'authentification des utilisateurs\n    - L'upload de fichiers audio de réunions\n    - La transcription automatique du contenu audio en texte\n    - La gestion et la consultation des transcriptions\n    \n    Développée avec FastAPI et intégrée avec AssemblyAI pour la transcription.\n    ","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Réunions"],"summary":"List Meetings","description":"Liste toutes les réunions de l'utilisateur connecté.\n\n- **status**: Filtre optionnel pour afficher uniquement les réunions avec un statut spécifique\n\nRetourne une liste de réunions avec leurs métadonnées (sans le contenu complet des transcriptions).","operationId":"list_meetings__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filtrer par statut de transcription (pending, processing, completed, error)","title":"Status"},"description":"Filtrer par statut de transcription (pending, processing, completed, error)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"},"title":"Response List Meetings  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Statut"],"summary":"Health Check","description":"Vérifie l'état de santé de l'API.\n\nCette route permet de vérifier si l'API est en ligne.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/register":{"post":{"tags":["Authentication","Authentication"],"summary":"Register","description":"Enregistre un nouvel utilisateur.\n\n- **email**: Adresse email valide\n- **password**: Mot de passe (minimum 8 caractères)\n- **full_name**: Nom complet\n\nRetourne les informations de l'utilisateur créé sans le mot de passe.\n\nExemple de réponse:\n```json\n{\n  \"message\": \"Utilisateur créé avec succès\",\n  \"user\": {\n    \"id\": \"550e8400-e29b-41d4-a716-446655440000\",\n    \"email\": \"utilisateur@example.com\",\n    \"full_name\": \"Nom Complet\",\n    \"created_at\": \"2024-03-01T14:30:45.123456\"\n  }\n}\n```","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate","description":"Informations de l'utilisateur à créer"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Register Auth Register Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication","Authentication"],"summary":"Login","description":"Authentifie un utilisateur et génère un token JWT.\n\n- **username**: Adresse email\n- **password**: Mot de passe\n\nRetourne un token d'accès JWT avec sa durée de validité.\n\nExemple de réponse:\n```json\n{\n  \"access_token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\",\n  \"token_type\": \"bearer\",\n  \"expires_in\": 1800\n}\n```\n\nUtilisez ce token dans l'en-tête d'autorisation pour accéder aux endpoints protégés:\n```\nAuthorization: Bearer {access_token}\n```","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Login Auth Login Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication","Authentication"],"summary":"Get Current User Info","description":"Récupère les informations de l'utilisateur actuellement connecté.\n\nCette route nécessite un token JWT valide et retourne les informations\nassociées à l'utilisateur authentifié.\n\nExemple de réponse:\n```json\n{\n  \"id\": \"550e8400-e29b-41d4-a716-446655440000\",\n  \"email\": \"utilisateur@example.com\",\n  \"full_name\": \"Nom Complet\",\n  \"created_at\": \"2024-03-01T14:30:45.123456\"\n}\n```","operationId":"get_current_user_info_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Current User Info Auth Me Get"}}}}},"security":[{"bearerAuth":[]}]}},"/upload":{"post":{"tags":["Réunions"],"summary":"Upload Meeting","description":"Télécharge un fichier audio MP3 et crée une nouvelle réunion avec transcription.\n\n- **file**: Fichier audio au format MP3 uniquement\n- **title**: Titre optionnel de la réunion (utilisera le nom du fichier par défaut)\n\nLe processus se déroule en plusieurs étapes:\n1. Upload du fichier\n2. Création d'une entrée dans la base de données\n3. Démarrage asynchrone de la transcription via AssemblyAI\n\nLa transcription peut prendre du temps en fonction de la durée de l'audio.","operationId":"upload_meeting_upload_post","security":[{"bearerAuth":[]}],"parameters":[{"name":"title","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_meeting_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Upload Meeting Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{meeting_id}":{"get":{"tags":["Réunions"],"summary":"Get Meeting Route","description":"Récupère les détails d'une réunion spécifique, y compris sa transcription.\n\n- **meeting_id**: Identifiant unique de la réunion\n\nRetourne toutes les informations de la réunion, y compris le texte de transcription\nsi la transcription est terminée.","operationId":"get_meeting_route__meeting_id__get","security":[{"bearerAuth":[]}],"parameters":[{"name":"meeting_id","in":"path","required":true,"schema":{"type":"string","description":"ID unique de la réunion","title":"Meeting Id"},"description":"ID unique de la réunion"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Meeting Route  Meeting Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Réunions"],"summary":"Update Meeting Route","description":"Met à jour les métadonnées d'une réunion.\n\n- **meeting_id**: Identifiant unique de la réunion\n- **meeting_update**: Données à mettre à jour (titre, notes, etc.)\n\nSeuls les champs non-nuls dans meeting_update seront modifiés.","operationId":"update_meeting_route__meeting_id__put","security":[{"bearerAuth":[]}],"parameters":[{"name":"meeting_id","in":"path","required":true,"schema":{"type":"string","description":"ID unique de la réunion","title":"Meeting Id"},"description":"ID unique de la réunion"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeetingUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Update Meeting Route  Meeting Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Réunions"],"summary":"Delete Meeting Route","description":"Supprime une réunion et ses données associées.\n\n- **meeting_id**: Identifiant unique de la réunion\n\nCette opération supprime à la fois les métadonnées de la réunion dans la base\nde données et le fichier audio associé s'il est stocké localement.","operationId":"delete_meeting_route__meeting_id__delete","security":[{"bearerAuth":[]}],"parameters":[{"name":"meeting_id","in":"path","required":true,"schema":{"type":"string","description":"ID unique de la réunion","title":"Meeting Id"},"description":"ID unique de la réunion"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Meeting Route  Meeting Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{meeting_id}/transcribe":{"post":{"tags":["Transcription"],"summary":"Transcribe Meeting Route","description":"Relance la transcription d'une réunion.\n\n- **meeting_id**: Identifiant unique de la réunion\n\nUtile si la transcription initiale a échoué ou pour relancer une transcription\navec de nouveaux paramètres (non implémenté actuellement).\n\nLa transcription s'exécute de manière asynchrone et peut prendre du temps\nselon la durée de l'audio.","operationId":"transcribe_meeting_route__meeting_id__transcribe_post","security":[{"bearerAuth":[]}],"parameters":[{"name":"meeting_id","in":"path","required":true,"schema":{"type":"string","description":"ID unique de la réunion","title":"Meeting Id"},"description":"ID unique de la réunion"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Transcribe Meeting Route  Meeting Id  Transcribe Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/{meeting_id}/transcript":{"get":{"tags":["Transcription"],"summary":"Get Transcript","description":"Récupère uniquement la transcription d'une réunion.\n\n- **meeting_id**: Identifiant unique de la réunion\n\nCette route est optimisée pour récupérer uniquement le texte de transcription\net son statut, sans les autres métadonnées de la réunion.","operationId":"get_transcript__meeting_id__transcript_get","security":[{"bearerAuth":[]}],"parameters":[{"name":"meeting_id","in":"path","required":true,"schema":{"type":"string","description":"ID unique de la réunion","title":"Meeting Id"},"description":"ID unique de la réunion"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Get Transcript  Meeting Id  Transcript Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Body_upload_meeting_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"Fichier audio MP3 à transcrire"}},"type":"object","required":["file"],"title":"Body_upload_meeting_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MeetingUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"transcript_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Transcript Text"},"transcript_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Transcript Status"}},"type":"object","title":"MeetingUpdate"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}